there's some sort of incorrect noise thing going on

+++ noise.h
@@ -21,7 +21,7 @@ public:
 	void setup(void);
 	double lerp(double a, double b, double t) const;
 
-	int valueTable[K_TABLE_SIZE];
+	double valueTable[K_TABLE_SIZE];
 	const unsigned char perm[K_TABLE_SIZE] = {
     225,155,210,108,175,199,221,144,203,116, 70,213, 69,158, 33,252,
     5, 82,173,133,222,139,174, 27,  9, 71, 90,246, 75,130, 91,191,
